+++ CEPROZAC/database/migrations/2018_02_25_220733_AlmacenLimpiezaMigrate.php
@@ -17,10 +17,10 @@ class AlmacenLimpiezaMigrate extends Migration
          $table->string('codigo')->nullable();
          $table->string('imagen')->nullable();
          $table->string('descripcion')->nullable();
-         $table->integer('cantidad');
+         $table->double('cantidad');
          $table->string('medida');
          $table->string('estado');
-          $table->integer('stock_minimo')->nullable();
+          $table->double('stock_minimo')->nullable();
          $table->timestamps();
      });
     }
