+++ CEPROZAC/database/migrations/2018_02_25_232132_AlmacenMaterialMigration.php
@@ -19,10 +19,10 @@ class AlmacenMaterialMigration extends Migration
             $table->string('codigo')->nullable();
             $table->string('imagen')->nullable();
             $table->string('descripcion')->nullable();
-            $table->integer('cantidad');
+            $table->double('cantidad');
             $table->string('medida')->nullable();
             $table->string('estado');
-             $table->integer('stock_minimo')->nullable();
+             $table->double('stock_minimo')->nullable();
                 $table->string('ubicacion')->nullable();
             $table->timestamps();
         });
